Start corrections to bootstrap result for vcv convergence
The initial 'proof' of theorem 3--convergence of the bootstrap distribution
to the limiting distribution of the statistic--is pretty wrong.  I'm
fixing it by explicitly invoking de Jong and Davidson's vcv result for the
convergence of the bootstrap variance.  This commit adds part of it, but
the LLN for the deriviatives and stochastic equicontinuity still need to be
established.