Add structural break to MC
Change the first monte carlo to add a structural break set of simulation; they
reflect really well on my statistic.  I also removed the second set of simulations
from the paper (the code is still there, but it's off the dependency path for the
pdf paper) since the first sims are so awesome.