Update multiple testing in empirical section
This commit changes the empirical section to use a much simpler
method of correcting for multiple comparisons. Instead of the
bootstrap we used before, we just simulate from the asymptotic
distribution and take the 1-alpha quantile of the largest
t-statistic. This correction makes the math and the presentation
a lot simpler, and complies with the editor's suggestion to
remove the bootstrap from the paper :).